class path cleaned

+++ Masterprojekt/src/GraphGenerators/Edge.java
@@ -1,38 +1,40 @@
 package GraphGenerators;
 
 public class Edge {
-   
-  Vertex from;
-  Vertex to;
-  int weight;
-  
-  
-  
-public Edge(Vertex from, Vertex to, int weight) {
-	super();
-	this.from = from;
-	this.to = to;
-	this.weight = weight;
-}
-public Vertex getFrom() {
-	return from;
-}
-public void setFrom(Vertex from) {
-	this.from = from;
-}
-public Vertex getTo() {
-	return to;
-}
-public void setTo(Vertex to) {
-	this.to = to;
-}
-public int getWeight() {
-	return weight;
-}
-public void setWeight(int weight) {
-	this.weight = weight;
-}
-  
-  
-	
+
+	Vertex from;
+	Vertex to;
+	double weight;
+
+	public Edge(Vertex from, Vertex to, int weight) {
+		super();
+		this.from = from;
+		this.to = to;
+		this.weight = weight;
+	}
+
+	public Vertex getFrom() {
+		return from;
+	}
+
+	public void setFrom(Vertex from) {
+		this.from = from;
+	}
+
+	public Vertex getTo() {
+		return to;
+	}
+
+	public void setTo(Vertex to) {
+		this.to = to;
+	}
+
+	public double getWeight() {
+		return weight;
+	}
+
+	public void setWeight(double weight) {
+		this.weight = weight;
+	}
+
 }
